Refactor annotation label editing and add tooltip UI
Moved the editable label widget to a reusable component (EditableLabel) and integrated it into annotation list and annotation widgets. Added a tooltip UI for editing annotation labels using the super_tooltip package. Updated annotation class ID change logic to handle invalid classes and finalize annotation updates. Added a global key for the image board and updated dependencies.
