+++ frontend/lib/utils/globals.dart
@@ -1,10 +1,13 @@
 import 'package:file_selector/file_selector.dart';
+import 'package:flutter/material.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 
 class Globals {
   Globals._();
 
   static String appVersion = '';
+
+  static GlobalKey globalImageBoardKey = GlobalKey();
 
   static const XTypeGroup imageType = XTypeGroup(
     label: 'images',
